% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldr.R
\name{world.f0}
\alias{world.f0}
\title{Extract fundamental frequencies (F0) using World}
\usage{
world.f0(
  w,
  frameshift = 5,
  f0floor = 71,
  f0ceil = 800,
  allowed_range = 0.1,
  method = "dio"
)
}
\arguments{
\item{frameshift}{Value of frame shift given in ms.}

\item{f0floor}{The lower limit of F0 (Hz).}

\item{allowed_range}{The allowed range}

\item{wave}{A Wave object to analyze.}

\item{f0}{A vector of F0 sequence. If NULL, the F0 is calculated in the function.}
}
\value{
A vector of F0 sequence.
}
\description{
\code{world.f0} extracts the fundamental frequency (F0) sequence using the World's standard F0 extraction algorithms, Dio and StoneMask.
If the Wave object has two or more channels, only the first channel (i.e. the left channel) will be used.
}
