% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldr.R
\name{world.stretch.spectrum}
\alias{world.stretch.spectrum}
\title{Stretch/shring the spectrum of the speech signal}
\usage{
world.stretch.spectrum(world, rate)
}
\arguments{
\item{world}{A list calculated by \code{world.analysis()}.}

\item{rate}{A float value for stretching}
}
\value{
A list in which the spectrum is changed.
}
\description{
\code{world.stretch.spectrum} changes the spectrum of the speech without changing the F0.
}
