% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldr.R
\name{world.stretch.time}
\alias{world.stretch.time}
\title{Stretch/shring the tempo of the speech signal}
\usage{
world.stretch.time(world, rate)
}
\arguments{
\item{world}{A list calculated by \code{world.analysis()}.}

\item{rate}{A float value for stretching}
}
\value{
A list in which the tempo is changed.
}
\description{
\code{world.stretch.time} changes the tempo of the speech without changing the F0.
}
