% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldr.R
\name{world.shift.f0}
\alias{world.shift.f0}
\title{Shift F0 values of the speech signal}
\usage{
world.shift.f0(world, rate)
}
\arguments{
\item{world}{A list calculated by \code{world.analysis()}.}

\item{rate}{A float value to be multiplied to the F0.}
}
\value{
A list in which the F0 values are changed.
}
\description{
\code{world.shift.f0} changes the F0 obtained by \code{world.analysis()}.
}
