% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldr.R
\name{world.synthesis}
\alias{world.synthesis}
\title{Synthesize speech signal using World}
\usage{
world.synthesis(world, normalize = FALSE)
}
\arguments{
\item{world}{A list calculated by \code{world.analysis()}.}

\item{normalize}{if TRUE, the output wave is normalized so that the maximum amplitude to be 32767.}
}
\value{
An Wave object, having one channel, PCM format. The sampling rate of the output will be the same as that of the input signal to the \code{world.analysis()}.
}
\description{
\code{world.synthesis} synthesizes the speech signal from the F0, spectrum and aperiodicity obtained by \code{world.analysis()}.
}
